rename_mq_producer_package
