+++ cdc/sink/mq_producer/kafka_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mqProducer
+package mq_producer
 
 import (
 	"testing"
